feat: refactored headers and added more styling
