+++ client/src/__tests__/components/Header.test.tsx
@@ -2,7 +2,7 @@ import { shallow } from 'enzyme';
 import Link from 'next/link';
 import * as React from 'react';
 
-import Header from './../../components/Header';
+import Header from './../../components/HeaderDefault';
 
 describe('<Header />', () => {
   it('renders without crashing', () => {
@@ -18,9 +18,9 @@ describe('<Header />', () => {
 
     it('contains link to /', () => {
       expect(
-        wrapper.find('Link').at(0).equals(
+        wrapper.contains(
           <Link href="/">
-            <a>Twitch-Stocks</a>
+            <a>Twitch Stocks</a>
           </Link>,
         ),
       ).toBe(true);
@@ -29,7 +29,7 @@ describe('<Header />', () => {
     // Temporarily set to '/' for now until we have auth
     it('contains link to /logout', () => {
       expect(
-        wrapper.find('Link').at(1).equals(
+        wrapper.contains(
           <Link href="/">
             <a>Logout</a>
           </Link>,
